refactor(figaresolver.utils.ts): refactor
transformNodes() method moved to figmaResolver. transformNodes() is the primary function that
handles mapping of functions for tokenGroups
